Textureに関する部分を別クラスに分離
