+++ Assets/MeshDeleterWithTexture/Editor/Utilities/TextureUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa79d3caf0914ab4b9f4bf091060c3cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
